WIP: Initial implementation of test base
